CAN-75: Change Password: Fronted API implementation: Fixes
 CAN-75: Change Password: Fronted API implementation: Fixes
